Image2Pixels.exe
Definitely read me for this, my project requires the installation of https://github.com/jacoborus/hexterm , so make sure you do that
Also! These are the steps when using my software:

1. Open Image2Pixels.exe
2. Select "Load Image" and find the image you would like
3. Press "Convert to Hex" button, it will take 30 mins depending on size of image, just wait until the cmd prompt has officially been exited automatically
4. Once the CMD prompt is exited, press "Convert to xterm-256", it should take around 3 minutes, do not press anything, check if
the program is still responding, but do not make it crash! so dont touch it for around 5 - 7 minutes to be safe.
5. Copy the code in the bottom left output box, that is your xterm-256 data!